# bpo-12029: Exception handling should match subclasses