--- a/Misc/NEWS.d/next/Core and Builtins/2018-04-13-10-38-00.bpo-12029.851NV8.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2018-04-13-10-38-00.bpo-12029.851NV8.rst
@@ -1,0 +1,3 @@
+Fix bug where exception matching fails to match virtual subclasses because
+``PyErr_GivenExceptionMatches`` used ``PyType_IsSubtype`` instead of
+``PyObject_IsSubclass``.