@mbmccoy 2018-05-19
Ok, I've run a bench mark to test the performance hit. For the benchmark code below that tests the "long path" in exception handling, I'm seeing a 1.4% slowdown over 100 million runs. To be precise, I'm seeing

| master | bpo-12029 (this branch) | % change| 
|-|-|-|
| 0.533µs | 0.540µs| 1.4%|

Here's a graph that batches the runs into 100000 trial each and shows the variability in runs:

<img width="400" alt="slowdown" src="https://user-images.githubusercontent.com/8784746/40269775-8550402c-5b37-11e8-8101-c25756f2a807.png">

The slowdown, while real, is not particularly significant: There are many cases where the code in `master` takes longer to run due simply to OS noise. 

You can reproduce this analysis via the code:


```python
import time
import json

time_us = 0
num_runs = 1000
num_tests_per_run = 100000
stats = []

for n in range(num_runs):
    start_time = time.time()
    for i in range(num_tests_per_run):
        try:
            {}["a"]
        except TypeError:
            pass
        except ValueError:
            pass
        except KeyError:
            pass

    end_time = time.time()
    time_us = (end_time - start_time) / num_tests_per_run * 1e6
    stats.append(time_us)

    time.sleep(1e-4)  # Make less sensitive to initial OS state

print("Took {} µs per test".format(sum(stats)/len(stats)))

import os
filename = os.path.abspath("stats-{}.json".format(time.time()))
print("Writing to file {}".format(filename))
with open(filename, 'w') as fp:
    json.dump(stats, fp)
```